+++ src/editor/dataset/constant/Element.ts
@@ -26,7 +26,9 @@ export const EDITOR_ELEMENT_COPY_ATTR: Array<keyof IElement> = [
   'url',
   'hyperlinkId',
   'dateId',
-  'dateFormat'
+  'dateFormat',
+  'level',
+  'titleId'
 ]
 
 export const EDITOR_ELEMENT_ZIP_ATTR: Array<keyof IElement> = [
@@ -52,7 +54,8 @@ export const EDITOR_ELEMENT_ZIP_ATTR: Array<keyof IElement> = [
   'control',
   'checkbox',
   'dateFormat',
-  'block'
+  'block',
+  'level'
 ]
 
 export const TEXTLIKE_ELEMENT_TYPE: ElementType[] = [
@@ -62,4 +65,16 @@ export const TEXTLIKE_ELEMENT_TYPE: ElementType[] = [
   ElementType.SUPERSCRIPT,
   ElementType.CONTROL,
   ElementType.DATE
+]
+
+export const INLINE_ELEMENT_TYPE: ElementType[] = [
+  ElementType.BLOCK,
+  ElementType.PAGE_BREAK,
+  ElementType.SEPARATOR,
+  ElementType.TABLE
+]
+
+export const INLINE_NODE_NAME: string[] = [
+  'HR',
+  'TABLE'
 ]